% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{create_unified_columns}
\alias{create_unified_columns}
\title{create_unified_columns generic}
\usage{
create_unified_columns(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to or from other methods}
}
\description{
create_unified_columns generic
}
