% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{create_unified_columns.splfmt_rts_v1}
\alias{create_unified_columns.splfmt_rts_v1}
\title{create_unified_columns}
\usage{
\method{create_unified_columns}{splfmt_rts_v1}(x)
}
\arguments{
\item{x}{x}

\item{...}{Arguments passed to or from other methods}
}
\description{
create_unified_columns
}
