% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{expand_time_to}
\alias{expand_time_to}
\title{Expand time to}
\usage{
expand_time_to(
  x,
  max_isoyear = NULL,
  max_isoyearweek = NULL,
  max_date = NULL,
  ...
)
}
\arguments{
\item{x}{An object of type \code{\link{splfmt_rts_data_v1}}}

\item{max_isoyear}{Maximum isoyear}

\item{max_isoyearweek}{Maximum isoyearweek}

\item{max_date}{Maximum date}

\item{...}{Not used.}
}
\description{
Attempts to expand the dataset to include more time

A time series is defined as a unique combination of:
\itemize{
\item granularity_time
\item granularity_geo
\item country_iso3
\item location_code
\item border
\item age
\item sex
\item *_id
\item *_tag
}
}
\seealso{
Other splfmt_rts_data: 
\code{\link{create_unified_columns}()},
\code{\link{heal}()},
\code{\link{identify_data_structure}()},
\code{\link{remove_class_splfmt_rts_data}()},
\code{\link{set_splfmt_rts_data_v1}()},
\code{\link{unique_time_series}()}
}
\concept{splfmt_rts_data}
