% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate test data}
\usage{
generate_test_data(fmt = "splfmt_rts_data_v1")
}
\arguments{
\item{fmt}{Data format (\code{splfmt_rts_data_v1})}
}
\description{
Generates some test data
}
\examples{
spltidy::generate_test_data("splfmt_rts_data_v1")
}
