% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{hash_data_structure}
\alias{hash_data_structure}
\title{heal generic}
\usage{
hash_data_structure(x, var, ...)
}
\arguments{
\item{x}{An object}

\item{var}{variable to hash}

\item{...}{Arguments passed to or from other methods}
}
\description{
heal generic
}
