% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{hash_data_structure.tbl_Pool}
\alias{hash_data_structure.tbl_Pool}
\title{Hash}
\usage{
\method{hash_data_structure}{tbl_Pool}(x, var, ...)
}
\arguments{
\item{x}{x}

\item{var}{variable to hash}

\item{...}{Arguments passed to or from other methods}
}
\description{
Hash
}
