% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{heal_time_csfmt_rts_data_v1}
\alias{heal_time_csfmt_rts_data_v1}
\title{Heals time}
\usage{
heal_time_csfmt_rts_data_v1(x, cols, granularity_time = "date")
}
\arguments{
\item{x}{The timepoints that need healing}

\item{cols}{The columns that should be returned}

\item{granularity_time}{The granularity of x}
}
\description{
Heals time
}
