% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csfmt_rts_v2.R
\name{heal_time_csfmt_rts_data_v2}
\alias{heal_time_csfmt_rts_data_v2}
\title{Provides corresponding healed times}
\usage{
heal_time_csfmt_rts_data_v2(x, cols, granularity_time = "date")
}
\arguments{
\item{x}{A vector containing either dates, isoyearweek, or isoyear.}

\item{cols}{Columns to restrict the output to.}

\item{granularity_time}{date, isoyearweek, or isoyear, depending on the values contained in x.}
}
\value{
data.table, a dataset with time columns corresponding to the values given in x.
}
\description{
Provides corresponding healed times
}
