% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{pivot_wider.data.table}
\alias{pivot_wider.data.table}
\title{collect}
\usage{
\method{pivot_wider}{data.table}(
  data,
  id_cols = NULL,
  names_from = name,
  names_prefix = "",
  names_sep = "_",
  names_glue = NULL,
  names_sort = FALSE,
  names_repair = "check_unique",
  values_from = value,
  values_fill = NULL,
  values_fn = NULL,
  ...
)
}
\arguments{
\item{data}{tbl object}

\item{id_cols}{id_cols}

\item{names_from}{names_from}

\item{names_prefix}{names_prefix}

\item{names_sep}{names_sep}

\item{names_glue}{names_glue}

\item{names_sort}{names_sort}

\item{names_repair}{names_repair}

\item{values_from}{values_from}

\item{values_fill}{values_fill}

\item{values_fn}{values_fn}

\item{...}{dots}
}
\description{
collect
}
