% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{plot.splfmt_rts_data_structure_hash_v1}
\alias{plot.splfmt_rts_data_structure_hash_v1}
\title{print}
\usage{
\method{plot}{splfmt_rts_data_structure_hash_v1}(x, y, ...)
}
\arguments{
\item{x}{x}

\item{y}{x}

\item{...}{dots}
}
\description{
print
}
\examples{
x <- spltidy::test_data_generator() \%>\%
  spltidy::set_splfmt_rts_data_v1() \%>\%
  spltidy::hash_data_structure("deaths_n") \%>\%
  plot
}
