% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{plot_epicurve.splfmt_rts_data_v1}
\alias{plot_epicurve.splfmt_rts_data_v1}
\title{Epicurve}
\usage{
\method{plot_epicurve}{splfmt_rts_data_v1}(x, ...)
}
\arguments{
\item{x}{Dataset}

\item{...}{X}
}
\description{
Epicurve
}
\examples{
splstyle::plot_epicurve(spltidy::norway_covid19_cases_by_time_location[location_code == "county03"], type = "single", var_y = "covid19_cases_testdate_n")
}
