% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_spl.R
\name{read_cs}
\alias{read_cs}
\title{Reads an object in an encrypted file serialized to disk}
\usage{
read_cs(filepath, private_key_path = Sys.getenv("ENCRYPTR_ID_RSA"))
}
\arguments{
\item{filepath}{The filepath (must end in '.cs.encrypted' or '.cs')}

\item{private_key_path}{Path to private key}
}
\description{
Reads an object in an encrypted file serialized to disk
}
