% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_spl.R
\name{save_spl}
\alias{save_spl}
\title{Saves (serializes) an encrypted object to disk}
\usage{
save_spl(
  x,
  filepath,
  nthreads = 1,
  public_key_path = Sys.getenv("ENCRYPTR_ID_RSA_PUB")
)
}
\arguments{
\item{x}{The object to serialize}

\item{filepath}{The filepath (must end in '.spl.encrypted' or '.spl')}

\item{nthreads}{Number of threads to use. Default 1.}

\item{public_key_path}{Path to public key}
}
\description{
Saves (serializes) an encrypted object to disk
}
