% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{set_splfmt_rts_data_v1}
\alias{set_splfmt_rts_data_v1}
\title{Convert data.table to splfmt_rts_data_v1 by reference}
\usage{
set_splfmt_rts_data_v1(x, create_unified_columns = TRUE, heal = TRUE)
}
\arguments{
\item{x}{The data.table to be converted to splfmt_rts_data_v1}

\item{create_unified_columns}{Do you want it to \code{\link{create_unified_columns}}?}

\item{heal}{Do you want to \code{\link{heal}} on creation?}
}
\value{
An extended \code{data.table} which is modified by reference and returned (invisibly).
}
\description{
\code{set_splfmt_rts_data_v1} converts a \code{data.table} to \code{splfmt_rts_data_v1} by reference.
}
\details{
For more details see the vignette:
\code{vignette("splfmt_rts_data_v1", package = "spltidy")}
}
\examples{
d <- spltidy::test_data_generator(fmt = "splfmt_rts_data_v1")
spltidy::set_splfmt_rts_data_v1(d, create_unified_columns = TRUE)
d[1, isoyearweek := "2021-01"]
d
d[2, isoyear := 2019]
d
d[4:5, date := as.Date("2020-01-01")]
d
d[10, c("isoyear", "isoyearweek") := .(2021, "2021-01")]
d
d[10, c("location_code") := .("norge")]
d
}
