% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{set_splfmt_rts_v1}
\alias{set_splfmt_rts_v1}
\title{Set as splfmt_rts_v1}
\usage{
set_splfmt_rts_v1(x, create_unified_variables = TRUE, heal = TRUE)
}
\arguments{
\item{x}{The data.table to be converted to splfmt_rts_v1}

\item{create_unified_variables}{Do you want it to create all unified variables?}

\item{heal}{Do you want to heal on creation?}
}
\description{
Set as splfmt_rts_v1
}
\examples{
d <- test_data_generator()
set_splfmt_rts_v1(d, create_unified_variables = TRUE)
d[1, isoyearweek := "2021-01"]
d
d[2, isoyear := 2019]
d
d[4:5, date := as.Date("2020-01-01")]
d
d[10, c("isoyear", "isoyearweek") := .(2021, "2021-01")]
d
d[10, c("location_code") := .("norge")]
d
}
