% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{set_splfmt_rts_v1_mini}
\alias{set_splfmt_rts_v1_mini}
\title{Set as splfmt_rts_v1_mini}
\usage{
set_splfmt_rts_v1_mini(x)
}
\arguments{
\item{x}{The data.table to be converted to splfmt_rts_v1_mini}
}
\description{
Set as splfmt_rts_v1_mini
}
\examples{
d <- test_data_generator()
set_splfmt_rts_v1_mini(d)
d[1,isoyearweek := "2021-01"]
d
d[2,isoyear := 2019]
d
d[4:5,date := as.Date("2020-01-01")]
d
d[10,c("isoyear","isoyearweek") := .(2021,"2021-01")]
d
d[10,c("location_code") := .("norge")]
d
}
