% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{test_data_generator}
\alias{test_data_generator}
\title{Test data generator}
\usage{
test_data_generator(fmt = "splfmt_rts_v1_mini")
}
\arguments{
\item{fmt}{Format (splfmt_rts_v1_mini)}
}
\description{
Test data generator
}
\examples{
test_data_generator("splfmt_rts_v1_mini")
}
