% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{unique_time_series}
\alias{unique_time_series}
\title{Unique time series}
\usage{
unique_time_series(x, set_time_series_id = FALSE, ...)
}
\arguments{
\item{x}{An object of type \code{\link{splfmt_rts_data_v1}}}

\item{set_time_series_id}{If TRUE, then \code{x} will have a new column called 'time_series_id'}

\item{...}{Not used.}
}
\description{
Attempts to identify the unique time series that exist in this dataset.

A time series is defined as a unique combination of:
\itemize{
\item granularity_time
\item granularity_geo
\item country_iso3
\item location_code
\item border
\item age
\item sex
\item *_id
\item *_tag
}
}
\seealso{
Other splfmt_rts_data: 
\code{\link{create_unified_columns}()},
\code{\link{heal}()},
\code{\link{identify_data_structure}()},
\code{\link{remove_class_splfmt_rts_data}()},
\code{\link{set_splfmt_rts_data_v1}()}
}
\concept{splfmt_rts_data}
