% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csfmt_rts_generics.R
\name{generate_test_data}
\alias{generate_test_data}
\title{Generate test data}
\usage{
generate_test_data(fmt = "csfmt_rts_data_v2")
}
\arguments{
\item{fmt}{Data format (\code{csfmt_rts_data_v2})}
}
\value{
csfmt_rts_data_v2, a dataset containing fake data.
}
\description{
Generates some test data
}
\examples{
cstidy::generate_test_data("csfmt_rts_data_v2")
}
