% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{hash_data_structure}
\alias{hash_data_structure}
\title{heal generic}
\usage{
hash_data_structure(x, var, ...)
}
\arguments{
\item{x}{An object}

\item{var}{variable to hash}

\item{...}{Arguments passed to or from other methods}
}
\description{
heal generic
}
\examples{
x <- spltidy::test_data_generator() \%>\%
  spltidy::set_splfmt_rts_data_v1() \%>\%
  spltidy::hash_data_structure("deaths_n") \%>\%
  plot
}
\seealso{
Other splfmt_rts_data: 
\code{\link{create_unified_columns}()},
\code{\link{heal}()},
\code{\link{remove_class_splfmt_rts_data}()},
\code{\link{set_splfmt_rts_data_v1}()}
}
\concept{splfmt_rts_data}
