% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{hash_data_structure}
\alias{hash_data_structure}
\title{Hash the data structure of a dataset for a given column}
\usage{
hash_data_structure(x, col, ...)
}
\arguments{
\item{x}{An object}

\item{col}{Column name to hash}

\item{...}{Arguments passed to or from other methods}
}
\description{
Reduces the data structure of a column inside a dataset into something that describes
}
\examples{
spltidy::generate_test_data() \%>\%
  spltidy::set_splfmt_rts_data_v1() \%>\%
  spltidy::hash_data_structure("deaths_n") \%>\%
  plot()
}
\seealso{
Other splfmt_rts_data: 
\code{\link{create_unified_columns}()},
\code{\link{heal}()},
\code{\link{remove_class_splfmt_rts_data}()},
\code{\link{set_splfmt_rts_data_v1}()}
}
\concept{splfmt_rts_data}
