% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{heal}
\alias{heal}
\title{Impute missing values}
\usage{
heal(x, ...)
}
\arguments{
\item{x}{An object of type \code{\link{splfmt_rts_data_v1}}}

\item{...}{Not used.}
}
\description{
Tries to impute missing values by performing smart assignment on all columns that are missing data.
E.g. if \code{location_code='norge'} then we know that \code{granularity_geo='nation'}.
}
\section{splfmt_rts_data_v1}{

The \strong{columns in bold} will be used to impute the listed columns.

\strong{location_code}:
\itemize{
\item granularity_geo
\item country_iso3
}

\strong{isoyear} (when \code{granularity_time=="isoyear"}):
\itemize{
\item isoweek
\item isoyearweek
\item season
\item seasonweek
\item calyear
\item calmonth
\item calyearmonth
\item date
}

\strong{isoyearweek} (when \code{granularity_time=="isoweek"}):
\itemize{
\item granularity_time
\item isoyear
\item isoweek
\item season
\item seasonweek
\item calyear
\item calmonth
\item calyearmonth
\item date
}

\strong{date} (when \code{granularity_time=="day"}):
\itemize{
\item granularity_time
\item isoyear
\item isoweek
\item isoyearweek
\item season
\item seasonweek
\item calyear
\item calmonth
\item calyearmonth
}

With regards to the time columns, \code{granularity_time} takes precedence over everything.
If \code{granularity_time} is missing, then we try to impute \code{granularity_time} by seeing if
there is only one time column with non-missing data. Due to the multitude of time columns,
\code{granularity_time} is an extremely important column and should always be kept with valid values.
}

\seealso{
Other splfmt_rts_data: 
\code{\link{create_unified_columns}()},
\code{\link{hash_data_structure}()},
\code{\link{remove_class_splfmt_rts_data}()},
\code{\link{set_splfmt_rts_data_v1}()}
}
\concept{splfmt_rts_data}
