% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{identify_data_structure}
\alias{identify_data_structure}
\title{Hash the data structure of a dataset for a given column}
\usage{
identify_data_structure(x, col, ...)
}
\arguments{
\item{x}{An object}

\item{col}{Column name to hash}

\item{...}{Arguments passed to or from other methods}
}
\description{
Reduces the data structure of a column inside a dataset into something that describes
}
\examples{
cstidy::generate_test_data() \%>\%
  cstidy::set_splfmt_rts_data_v1() \%>\%
  cstidy::identify_data_structure("deaths_n") \%>\%
  plot()
}
\seealso{
Other splfmt_rts_data: 
\code{\link{create_unified_columns}()},
\code{\link{expand_time_to}()},
\code{\link{heal}()},
\code{\link{remove_class_splfmt_rts_data}()},
\code{\link{set_splfmt_rts_data_v1}()},
\code{\link{unique_time_series}()}
}
\concept{splfmt_rts_data}
