% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csfmt_rts_v1.R
\name{identify_data_structure}
\alias{identify_data_structure}
\alias{identify_data_structure.csfmt_rts_data_v1}
\alias{identify_data_structure.tbl_Microsoft SQL Server}
\title{Hash the data structure of a dataset for a given column}
\usage{
identify_data_structure(x, col, ...)

\method{identify_data_structure}{csfmt_rts_data_v1}(x, col, ...)

\method{identify_data_structure}{`tbl_Microsoft SQL Server`}(x, col, ...)
}
\arguments{
\item{x}{An object}

\item{col}{Column name to hash}

\item{...}{Arguments passed to or from other methods}
}
\value{
csfmt_rts_data_structure_hash_v1, a summary object.
}
\description{
Reduces the data structure of a column inside a dataset into something that describes
}
\examples{
cstidy::generate_test_data() \%>\%
  cstidy::set_csfmt_rts_data_v1() \%>\%
  cstidy::identify_data_structure("deaths_n") \%>\%
  plot()
}
\seealso{
Other csfmt_rts_data: 
\code{\link{expand_time_to}()},
\code{\link{remove_class_csfmt_rts_data}()},
\code{\link{set_csfmt_rts_data_v1}()},
\code{\link{unique_time_series}()}
}
\concept{csfmt_rts_data}
