% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{pivot_wider.data.table}
\alias{pivot_wider.data.table}
\title{pivot_wider}
\usage{
\method{pivot_wider}{data.table}(data, ...)
}
\arguments{
\item{data}{tbl object}

\item{...}{dots}
}
\description{
pivot_wider
}
