% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csfmt_rts_v1.R
\name{remove_class_csfmt_rts_data}
\alias{remove_class_csfmt_rts_data}
\title{Remove class csfmt_rts_data_*}
\usage{
remove_class_csfmt_rts_data(x)
}
\arguments{
\item{x}{data.table}
}
\description{
Remove class csfmt_rts_data_*
}
\examples{
x <- cstidy::generate_test_data() \%>\%
  cstidy::set_csfmt_rts_data_v1()
class(x)
cstidy::remove_class_csfmt_rts_data(x)
class(x)
}
\seealso{
Other csfmt_rts_data: 
\code{\link{create_unified_columns}()},
\code{\link{expand_time_to}()},
\code{\link{heal}()},
\code{\link{identify_data_structure}()},
\code{\link{set_csfmt_rts_data_v1}()},
\code{\link{unique_time_series}()}
}
\concept{csfmt_rts_data}
