% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{remove_class_splfmt_rts_data}
\alias{remove_class_splfmt_rts_data}
\title{remove class splfmt_rts_data_*}
\usage{
remove_class_splfmt_rts_data(x)
}
\arguments{
\item{x}{data.table}
}
\description{
remove class splfmt_rts_data_*
}
