% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splfmt_rts_v1.R
\name{remove_class_splfmt_rts_data}
\alias{remove_class_splfmt_rts_data}
\title{Remove class splfmt_rts_data_*}
\usage{
remove_class_splfmt_rts_data(x)
}
\arguments{
\item{x}{data.table}
}
\description{
Remove class splfmt_rts_data_*
}
\examples{
x <- spltidy::generate_test_data() \%>\%
  spltidy::set_splfmt_rts_data_v1()
class(x)
spltidy::remove_class_splfmt_rts_data(x)
class(x)
}
\seealso{
Other splfmt_rts_data: 
\code{\link{create_unified_columns}()},
\code{\link{heal}()},
\code{\link{identify_data_structure}()},
\code{\link{set_splfmt_rts_data_v1}()},
\code{\link{unique_time_series}()}
}
\concept{splfmt_rts_data}
